Move xml stream join #201
